Make mpt_updates.rs file
